refactor: extract fragments and mutations
